+++ SunEngine.Commons/Controllers/BaseController.cs
@@ -6,8 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using SunEngine.Commons.Cache.CacheModels;
 using SunEngine.Commons.Cache.Services;
+using SunEngine.Commons.Errors;
 using SunEngine.Commons.Managers;
-using SunEngine.Commons.Misc;
 using SunEngine.Commons.Models;
 using SunEngine.Commons.Security;
 using SunEngine.Commons.Utils;
