+++ SunEngine.Commons/Controllers/BlogController.cs
@@ -7,7 +7,6 @@ using Microsoft.Extensions.Options;
 using SunEngine.Commons.Cache.CacheModels;
 using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.Configuration.Options;
-using SunEngine.Commons.Misc;
 using SunEngine.Commons.Presenters;
 using SunEngine.Commons.Security;
 using SunEngine.Commons.Utils.PagedList;
