+++ SunEngine.Commons/Managers/AccountManager.cs
@@ -9,9 +9,8 @@ using Flurl;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using SunEngine.Commons.Configuration.Options;
-using SunEngine.Commons.Controllers;
 using SunEngine.Commons.DataBase;
-using SunEngine.Commons.Misc;
+using SunEngine.Commons.Errors;
 using SunEngine.Commons.Models;
 using SunEngine.Commons.Services;
 using SunEngine.Commons.Utils;
