+++ SunEngine/Startup.cs
@@ -9,7 +9,7 @@ using Newtonsoft.Json;
 using SunEngine.Admin;
 using SunEngine.Commons.Configuration.AddServices;
 using SunEngine.Commons.DataBase;
-using SunEngine.Commons.Misc;
+using SunEngine.Commons.Errors;
 using SunEngine.Commons.Security;
 using SunEngine.Commons.Services;
 using SunEngine.Commons.Utils;
